requirements.txt
curses
github3.py
pyautogui
